refactor(realtime): split the transport from the connection hub
The server owned both the Bun socket handling and every piece of realtime
state, so nothing could drive it from another transport. Move the
connections map, presence, authentication, Redis fan-out and heartbeat
into a hub that accepts an abstract socket, and leave Bun.serve as a thin
adapter over it. A second adapter can now attach a different socket
implementation to the same hub.
